server_IA/requirements.txt
flask
requests
flask-limiter
python-dotenv
gunicorn